[Refactor] Extract shared CP symmetric-memory infrastructure
